fix: PR review fixes — rename overlay, remove dead code, fix height measurement
- Rename save-template-overlay.tsx to export-overlay.tsx to match component name
- Remove unused demo-gallery components (preserved on feat/demo-gallery branch)
- Fix widget height measurement using offscreen clone instead of collapsing
  visible DOM, preventing potential ResizeObserver infinite loops
- Use camelCase keyElements prop in PlanCard (remap from snake_case at call site)
- Update chat disclaimer to be relevant to visualization demo
