+++ apps/app/src/components/generative-ui/charts/pie-chart.tsx
@@ -6,7 +6,7 @@ import {
 } from "recharts";
 import { z } from "zod";
 import { CHART_COLORS, CHART_CONFIG } from "./config";
-import { ExportOverlay } from "../save-template-overlay";
+import { ExportOverlay } from "../export-overlay";
 
 export const PieChartProps = z.object({
   title: z.string().describe("Chart title"),
